Sample screenshot
If the user chooses to view functions or aliases instead, the interface is very similar. In this case however, the user requested a specific alias, rather than relying on a menu of choices and the contextual view therefore demonstrates that the command requested is an alias, which function it calls, the content of that function and what the set-alias command would look like in order to create the alias in question.